requirements.txt
python-telegram-bot
apscheduler
facebook-sdk